requirements.txt
openai~=1.7.0
tiktoken~=0.4.0
python-dotenv~=1.0.0